fixed issue12
now one can see a progress bar when the actual registration (which means when elastix is starting) begins. that elastix registration takes most of the time and thats why the progress bar starts there and not right from the beginning because the first steps (e.g. the creation of the dapimask) is rather fast. and also it would be really complicated to create a progress bar right from the beginning :D